% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacoxph.R
\name{bacoxph.fit}
\alias{bacoxph.fit}
\title{Internal fit for bacoxph}
\usage{
bacoxph.fit(
  x,
  y,
  offset = rep(0, nobs),
  weights = rep(1, nobs),
  init = 0,
  strats = NULL,
  control = coxph.control(eps = 1e-04, iter.max = 50),
  ties = "efron",
  prior = "t",
  group = NULL,
  method.coef = NULL,
  prior.mean = 0,
  prior.sd = 1,
  prior.scale = 1,
  prior.df = 1,
  autoscale = TRUE,
  ss = c(0.05, 0.1),
  b = 1,
  theta.weights = NULL,
  inter.hierarchy = NULL,
  inter.parents = NULL,
  Warning = FALSE
)
}
\arguments{
\item{Warning}{}
}
\description{
Internal fit for bacoxph
}
\keyword{internal}
