% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bacoxph.R
\name{bacoxph}
\alias{bacoxph}
\title{Bayesian cox additive model}
\usage{
bacoxph(
  formula,
  data,
  weights,
  subset,
  na.action,
  init,
  control = coxph.control(eps = 1e-04, iter.max = 50),
  ties = c("breslow", "efron"),
  tt,
  prior = Student(0, 0.5, 1),
  group = NULL,
  method.coef,
  theta.weights = NULL,
  inter.hierarchy = NULL,
  inter.parents = NULL,
  Warning = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{formula, data, weights, subset, na.action, init, control, ties, tt}{These arguments are the same as in \code{\link{coxph}} in the package survival}

\item{prior, group, method.coef, verbose, theta.weights, inter.hierarchy, inter.parents}{These arguments are the same as in \code{\link{bgam}}}

\item{...}{Reserved for future arguments.}
}
\value{
This function returns an object of class "coxph.penal" and "coxph", including all outputs from the function coxph and also results for the additional parameters in the hierarchical models.
}
\description{
Bayesian cox additive model
}
