% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_selection.R
\name{bamlasso_var_selection}
\alias{bamlasso_var_selection}
\title{Variable Selection for Bamlasso}
\usage{
bamlasso_var_selection(mdl)
}
\arguments{
\item{mdl}{Model returned from Bamlasso}
}
\value{
A list of two consisting a character vector, named Parametric, and a data frame, named Non-parametric. The character vector contains the parametric variabels selected in the model if any. The dataframe, 3-by-p, contains the selected smoothing function, separated for linear components and nonlinear components
}
\description{
Variable Selection for Bamlasso
}
