% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgam.R
\name{bgam}
\alias{bgam}
\title{Bayesian hierarchical additive model}
\usage{
bgam(
  formula,
  family = gaussian,
  data,
  offset,
  weights,
  subset,
  na.action,
  start = NULL,
  etastart,
  mustart,
  control = stats::glm.control(epsilon = 1e-04, maxit = 50),
  prior = Student(),
  group = NULL,
  method.coef,
  theta.weights = NULL,
  inter.hierarchy = NULL,
  inter.parents = NULL,
  prior.sd = 0.5,
  dispersion = 1,
  Warning = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}
}
\value{
This function returns an object of class "glm", including all outputs from the function \code{\link{glm}}, and also results for the additional parameters in the hierarchical models.
}
\description{
Bayesian hierarchical additive model
}
