% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgam.R
\name{bglm_spline.fit}
\alias{bglm_spline.fit}
\title{Internal spline fit for bgam}
\usage{
bglm_spline.fit(
  x,
  y,
  weights = rep(1, nobs),
  start = NULL,
  etastart = NULL,
  mustart = NULL,
  offset = rep(0, nobs),
  family = gaussian(),
  control = glm.control(),
  intercept = TRUE,
  prior = "de",
  group = NULL,
  method.coef = 1,
  dispersion = 1,
  prior.mean = 0,
  prior.sd = 0.5,
  prior.scale = 1,
  prior.df = 1,
  autoscale = TRUE,
  ss = c(0.05, 0.1),
  b = 1,
  theta.weights = NULL,
  inter.hierarchy = NULL,
  inter.parents = NULL,
  Warning = FALSE
)
}
\arguments{
\item{Warning}{}
}
\description{
Internal spline fit for bgam
}
\keyword{internal}
