% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAM.R
\docType{package}
\name{BHAM}
\alias{BHAM}
\title{BHAM: A package to fit Bayesian Hierarchical Additive Models for High Dimensional Application}
\description{
// TODO: Update this page
The foo package provides three categories of important functions:
foo, bar and baz.
}
\section{Foo functions}{

The foo functions ...
}

