% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bamlasso.R
\name{bmlasso_spline.fit}
\alias{bmlasso_spline.fit}
\title{Internal spline fit function for bamlasso}
\usage{
bmlasso_spline.fit(
  x,
  y,
  family = "gaussian",
  offset = NULL,
  epsilon = 1e-04,
  maxit = 50,
  init = rep(0, ncol(x)),
  alpha = 1,
  ss = c(0.04, 0.5),
  b = 1,
  group = NULL,
  theta.weights = NULL,
  inter.hierarchy = NULL,
  inter.parents = NULL,
  Warning = FALSE
)
}
\arguments{
\item{Warning}{}
}
\description{
Internal spline fit function for bamlasso
}
\keyword{internal}
