% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{construct_smooth_data}
\alias{construct_smooth_data}
\title{Creating design matrix based on spline configuration}
\usage{
construct_smooth_data(sm_df, dat)
}
\arguments{
\item{sm_df}{a data frame that has three columns, _Func_, _Var_, _Args_ (case sensitive). See example}

\item{dat}{the raw data set}
}
\value{
a dataframe that contains the design matrix of
}
\description{
contruct_smooth_data creates the design matrix according to the spline function defined
in sm_df. The spline design matrix is created using smoothCon from the pacakge mgcv.
}
