% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{create_smooth_name}
\alias{create_smooth_name}
\title{Create Smooth Design Matrix Column Names}
\usage{
create_smooth_name(smth)
}
\arguments{
\item{smth}{the Smooth object, created from smoothCon of mgcv}
}
\value{
a vector of string, containing the column names of the spline design matrix
}
\description{
Internal function for creating names for Smooth design matrix
}
\examples{
 smth <- (
  mgcv::smoothCon(object = mgcv::s(x), data = data.frame(x = rnorm(100)),
            scale.penalty = TRUE,
            absorb.cons = TRUE,
            null.space.penalty = TRUE,
            diagonal.penalty = TRUE)
)[[1]]

BHAM:::create_smooth_name(smth)
}
