% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_bgam.R
\name{cv.bgam}
\alias{cv.bgam}
\title{Cross-validation & Tuning for \code{bgam} and \code{bamlasso}}
\usage{
cv.bgam(object, nfolds = 10, foldid = NULL, ncv = 1, s0 = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Cross-validation & Tuning for \code{bgam} and \code{bamlasso}
}
\examples{

}
