% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_bgam.R
\name{cv.gam.glm}
\alias{cv.gam.glm}
\title{Internal function for cross validation of bgam model}
\usage{
cv.gam.glm(
  object,
  nfolds = 10,
  foldid = NULL,
  ncv = 1,
  s0 = NULL,
  group = group,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Internal function for cross validation of bgam model
}
