% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{make_group}
\alias{make_group}
\title{Create grouping for spline design matrix}
\usage{
make_group(.names, null_group = TRUE, shared_null = FALSE)
}
\arguments{
\item{.names}{spline matrix names, always in the format "var_name.baseX". Directly from Construct_Smooth_Data}

\item{null_group}{A indicator if the null space are in its own group, i.e. if null space are penalized}

\item{shared_null}{A indicator if the null space have a shared indicator with the penalized space}
}
\value{

}
\description{
Create grouping for spline design matrix
}
