% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_ancillary_parameters}
\alias{summarize_ancillary_parameters}
\title{Summarize ancillary parameters}
\usage{
summarize_ancillary_parameters(fit)
}
\arguments{
\item{fit}{the fitted object from bglm_spline or bmlasso_spline}
}
\value{
a data frame contains summarized ancillary parameters for spline components of each variable
}
\description{
Summarize ancillary parameters
}
