% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updates.r
\name{update.scale.p.group}
\alias{update.scale.p.group}
\title{Title}
\usage{
\method{update}{scale.p.group}(prior = "mde", df = 1, b0, ss, theta, group.vars)
}
\arguments{
\item{group.vars}{}
}
\description{
Title
}
