% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BHAM.R
\docType{package}
\name{BHAM}
\alias{BHAM}
\title{BHAM: A package to fit Bayesian Hierarchical Additive Models for High Dimensional Application}
\description{
BHAM is a freely avaible R pakcage that implments Bayesian
hierarchical additive models for high-dimensional clinical and genomic data.
The package includes functions that generlized additive model, and Cox additive
model with the spike-and-slab LASSO prior. These functions implements scalable
and stable algorithms to estimate parameters. BHAM also provides utility
functions to construct additive models in high dimensional settings,
select optimal models, summarize bi-level variable selection results,
and visualize nonlinear effects. The package can facilitate flexible
modeling of large-scale molecular data, i.e. detecting succeptable
variables and inforing disease diagnostic and prognostic.
}
