% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_smooth_data.r
\name{construct_smooth_data}
\alias{construct_smooth_data}
\title{Create a design matrix based on spline configuration}
\usage{
construct_smooth_data(sm_df, dat)
}
\arguments{
\item{sm_df}{a data frame that has three columns, \emph{Func}, \emph{Var}, \emph{Args} (case sensitive). See example}

\item{dat}{the raw data set}
}
\value{
a list contains
\itemize{
\item{data}{the contructed design matrix, the intercept and outcome is not included}
\item{Smooth}{}
}
}
\description{
create_smooth_data creates a design matrix with spline functions.
A data frame containing custom spline functions can be supplied or
spline arguements can be supplied to
in sm_df. The spline design matrix is created using smoothCon from the pacakge mgcv.
}
\examples{

raw_dat <- sim_Bai(100, 5)$dat \%>\% data.frame

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)



construct_smooth_data(sm_df, raw_dat)


}
