% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_smooth_data.r
\name{create_smooth_name}
\alias{create_smooth_name}
\title{Internal name creation for smooth spline}
\usage{
create_smooth_name(smth)
}
\arguments{
\item{smth}{}
}
\description{
Internal name creation for smooth spline
}
\keyword{internal}
