% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_group.r
\name{make_group}
\alias{make_group}
\title{Create grouping for spline design matrix
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}}
\usage{
make_group(.names, penalize_null = TRUE, shared_null = TRUE)
}
\arguments{
\item{.names}{spline matrix names, always in the format "var_name.baseX". Directly from Construct_Smooth_Data
#param null_group A indicator if the null space are in its own group, i.e. if null space are penalized}

\item{penalize_null}{do we penalize the null space, i.e. do we fit the null space into group parameter in bglm group parameter.}

\item{shared_null}{A indicator if the null space have a shared indicator with the penalized space}
}
\value{
A vector of lists, where each element list contains variables that belong to the same group
}
\description{
Create grouping for spline design matrix
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{

raw_dat <- sim_Bai(100, 5)$dat |> data.frame()

sm_df <- data.frame(
 Var = setdiff(names(raw_dat), "y"),
 Func = "s",
 Args ="bs='cr', k=5"
)

dsn_mat <- construct_smooth_data(sm_df, raw_dat)$data

make_group(names(dsn_mat))
}
