% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_bai.R
\name{sim_Bai}
\alias{sim_Bai}
\title{Simulate Spline Data}
\usage{
sim_Bai(n, p, family = binomial(), dispersion = 1)
}
\arguments{
\item{n}{Integer, sample size}

\item{p}{Integer, number of predictors. The minimum number of predictors is default to 4. If user input p<4, p changes to 4 with a warning message}

\item{family}{Family object, one of `binomial` (default), `gaussian` , `poisson`. Please see \code{\link[stats]{family}} for more detail}

\item{dispersion}{Double, required for gaussian distribution, match to sd parameter.}
}
\value{
a list contains `dat`, `eta`, `mu`
\itemize{
    \item{dat}{ a numeric matrix, contains predictors `x1` to `xp` where p is replaced by the integer value. The outcome variable is named `y`}
    \item{eta}{the linear predictor of the simulated data}
    \item{mu}{the mean of the simulated data}

}
}
\description{

}
\examples{
# Binomial Outcome
sim_Bai(100, 4)

# Logistic Outcome
sim_Bai(100, 4, family = poisson)

# Gaussian Outcome with measurement error variance = 1
sim_Bai(100, 4, family = gaussian, dispersion = 1 )
}
