% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_bgam.R
\name{tune.bgam}
\alias{tune.bgam}
\alias{cv.bgam}
\title{Cross-Validation for Bayesian Additive Hierarchical Models}
\usage{
tune.bgam(
  object,
  nfolds = 10,
  foldid = NULL,
  ncv = 1,
  s0 = NULL,
  verbose = TRUE
)

cv.bgam(object, nfolds = 10, foldid = NULL, ncv = 1, s0 = NULL, verbose = TRUE)
}
\arguments{
\item{verbose}{logical. If \code{TRUE}, print out the computational time and progress.}
}
\description{
he function \code{\link[BhGLM]{cv.bh}} performs K-fold cross-validation and calculates cross-validated predictive measures for Bayesian hierarchical GLMs and Cox survival model, or for elastic net from the package \bold{glmnet}.
}
